Cortex Ruler alerts improvements
Adds a set of improvements on the Cortex Ruler alerts.

The addition of:

- `CortexRulerNotConnectedToAlertmanagers` to know whenever a ruler is
  connected to _any_ Alertmanager.
- `CortexRulerErrorSendingAlertsToAnyAlertmanager` to have a threshold for sending to at least _an Alertmanager.
- `CortexRulerErrorSendingAlertsToSomeAlertmanagers` to have a threshold
  for sending to each alertmanager.
- `CortexRulerNotificationQueueRunningFull` to know whenever the
  notification queue will be full.

On top of that, I have modified the existing rule evaluation alerts to
include in the aggregation the instance and rule group. Given that the
ruler shards by rule group this makes it easier to identify the
offending rules.

Signed-off-by: gotjosh <josue@grafana.com>