Cortex kv panels: narrow to one instance
Distributor uses multiple kv stores - for global limits and ha-tracker,
as well as reading from the ingester ring - so we need to narrow the
panel to just the one it says it is showing.

For consistency, do the same on the ingester panel, although currently
ingesters only have one kv store.

Note that the renamed recording rule will mean that dashboards show
no data for latency prior to the change.